DD-32-refactor: 수정사항 반영
