+++ src/pages/users/[user_id]/index.tsx
@@ -1,0 +1,9 @@
+import MyPageRegistered from '@/components/pageComponents/MyPage/MyPageRegistered';
+
+export default function MyPage() {
+  return (
+    <div>
+      <MyPageRegistered />
+    </div>
+  );
+}
